+++ __tests__/typescript.test.js
@@ -26,7 +26,7 @@ describe('TypeScript config', () => {
 			'@typescript-eslint/no-non-null-assertion',
 			'@typescript-eslint/no-unnecessary-condition',
 			'@typescript-eslint/prefer-optional-chain',
-			'import/no-duplicates',
+			'import-x/no-duplicates',
 			// Replaces `@typescript-eslint/no-unnecessary-boolean-literal-compare`,
 			// which we turn off in favour of this one
 			'unicorn/no-unnecessary-boolean-comparison',